requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.40.0
gradio>=4.0.0
Pillow>=10.0.0

# Model quantization
bitsandbytes>=0.41.0
accelerate>=0.20.0

# Utilities
pathlib
datetime
socket
threading
hashlib

# Optional but recommended
numpy>=1.24.0
scipy>=1.10.0